# docs/swagger/swagger.yaml
definitions:
  handler.CreateTaskRequest:
    properties:
      description:
        type: string
      priority:
        maximum: 3
        minimum: 1
        type: integer
      title:
        minLength: 3
        type: string
      user_id:
        type: string
    required:
    - priority
    - title
    - user_id
    type: object
  handler.CreateUserRequest:
    properties:
      email:
        type: string
      name:
        minLength: 3
        type: string
    required:
    - email
    - name
    type: object
  handler.TaskResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      priority:
        type: integer
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  handler.UpdateTaskRequest:
    properties:
      description:
        type: string
      priority:
        maximum: 3
        minimum: 1
        type: integer
      title:
        type: string
    required:
    - title
    type: object
  handler.UpdateTaskStatusRequest:
    properties:
      status:
        enum:
        - new
        - in_progress
        - completed
        type: string
    required:
    - status
    type: object
  handler.UpdateUserRequest:
    properties:
      email:
        type: string
      name:
        minLength: 3
        type: string
    required:
    - email
    - name
    type: object
  handler.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/tasks:
    post:
      consumes:
      - application/json
      description: Create a new task for a user
      parameters:
      - description: Task data
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/handler.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.TaskResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new task
      tags:
      - tasks
  /api/v1/tasks/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a task by ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Delete a task
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Update title, description or priority
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated data
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.TaskResponse'
      summary: Update a task
      tags:
      - tasks
  /api/v1/tasks/{id}/status:
    patch:
      consumes:
      - application/json
      description: Update the status of a specific task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Status data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateTaskStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.TaskResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update task status
      tags:
      - tasks
  /api/v1/tasks/{user_id}:
    get:
      consumes:
      - application/json
      description: List all tasks for a given user ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.TaskResponse'
            type: array
      summary: List tasks by user
      tags:
      - tasks
  /api/v1/users:
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new user
      tags:
      - users
  /api/v1/users/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Delete a user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get a single user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UserResponse'
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update name or email of a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UserResponse'
      summary: Update a user
      tags:
      - users
swagger: "2.0"